+++ src/blob/index/core.rs
@@ -59,14 +59,33 @@ struct MemoryAttrs<K> {
     marker: PhantomData<K>,
 }
 
+const BTREE_B_FACTOR: usize = 6;
+const BTREE_VALUES_LEN: usize = BTREE_B_FACTOR * 2 - 1;
+const BTREE_EDGES_LEN: usize = BTREE_B_FACTOR * 2;
+
 impl<K> MemoryAttrs<K>
 where
     for<'a> K: Key<'a>,
 {
-    const KEY_SIZE: usize = K::LEN as usize;
-    const BPTREE_ENTRY_SIZE: usize =
-        size_of::<Vec<u8>>() + MemoryAttrs::<K>::KEY_SIZE + size_of::<Vec<RecordHeader>>();
-    const RECORD_HEADER_SIZE: usize = size_of::<RecordHeader>() + MemoryAttrs::<K>::KEY_SIZE;
+    const BTREE_ENTRY_SIZE: usize = K::MEM_SIZE + size_of::<Vec<RecordHeader>>();
+    const RECORD_HEADER_SIZE: usize = size_of::<RecordHeader>() + K::LEN as usize;
+    // Each node in BTreeMap contains preallocated vectors of 11 values and 12 edges.
+    // Although count of nodes can't be determined without reimplementing insertion algorithm,
+    // we can use approximation of overhead size added per one key
+    const BTREE_DATA_NODE_SIZE: usize = size_of::<Option<std::ptr::NonNull<()>>>() +                     // ptr to parent
+                                        size_of::<u16>() * 2 +                                           // metadata
+                                        MemoryAttrs::<K>::BTREE_ENTRY_SIZE * BTREE_VALUES_LEN;           // data
+    const BTREE_DATA_NODE_RATIO: f64 = 1.0 / BTREE_VALUES_LEN as f64;
+    const BTREE_INTERNAL_NODE_OVERHEAD: usize = size_of::<std::ptr::NonNull<()>>() * BTREE_EDGES_LEN;    // edges
+    const BTREE_INTERNAL_NODE_RATIO: f64 = (1 +
+                                            BTREE_EDGES_LEN +
+                                            BTREE_EDGES_LEN.pow(2) +
+                                            BTREE_EDGES_LEN.pow(3) +
+                                            BTREE_EDGES_LEN.pow(4)) as f64
+                                           / (BTREE_VALUES_LEN * BTREE_EDGES_LEN.pow(5)) as f64;
+    const BTREE_SIZE_MULTIPLIER: f64 =
+        (MemoryAttrs::<K>::BTREE_DATA_NODE_SIZE as f64 * MemoryAttrs::<K>::BTREE_DATA_NODE_RATIO) +
+        (MemoryAttrs::<K>::BTREE_INTERNAL_NODE_OVERHEAD as f64 * MemoryAttrs::<K>::BTREE_INTERNAL_NODE_RATIO);
 }
 
 pub type InMemoryIndex<K> = BTreeMap<K, Vec<RecordHeader>>;
@@ -99,13 +118,13 @@ where
             ..
         } = &mem;
         let len = self.headers.len();
-        trace!("record_header_size: {}, records_allocated: {}, data.len(): {}, entry_size (key + vec): {}",
-                MemoryAttrs::<K>::RECORD_HEADER_SIZE, records_allocated, len, MemoryAttrs::<K>::BPTREE_ENTRY_SIZE);
+        trace!("len: {}, records_allocated: {}, records_count: {}",
+                len, records_allocated, records_count);
         // last minus is neccessary, because allocated but not initialized record
         // headers don't have key allocated on heap
         MemoryAttrs::<K>::RECORD_HEADER_SIZE * records_allocated
-            + len * MemoryAttrs::<K>::BPTREE_ENTRY_SIZE
-            - (records_allocated - records_count) * MemoryAttrs::<K>::KEY_SIZE
+            + (len as f64 * MemoryAttrs::<K>::BTREE_SIZE_MULTIPLIER) as usize
+            - (records_allocated - records_count) * K::LEN as usize
     }
 
     fn records_count(&self) -> usize {